fix: drop xargs from postStart script
xargs has inconsistant behaviour handling whitespace:
* When given argument `-I {}` it no longer splits on spaces properly.
* When given `-d" "` to force it to split on spaces it doesn't handle
  single values properly.

This problem identifies itself when there are multiple create models defined in
the values today because xargs tries to put every model name in a single
create command.

This commit drops the use of xargs completely from the script and uses
the helm templating to create all the commands explictly.